Rename: 조회용 dto 폴더명 변경
- scrap -> inquiry
